refactor: move each pooler into its own module
