+++ colpali_engine/compression/__init__.py
@@ -1,1 +1,6 @@
-from .token_pooling import BaseTokenPooler, HierarchicalTokenPooler, TokenPoolingOutput
+from .token_pooling import (
+    BaseTokenPooler,
+    HierarchicalTokenPooler,
+    LambdaTokenPooler,
+    TokenPoolingOutput,
+)
