+++ colpali_engine/compression/token_pooling/__init__.py
@@ -1,0 +1,3 @@
+from .base_token_pooling import BaseTokenPooler, TokenPoolingOutput
+from .hierarchical_token_pooling import HierarchicalTokenPooler
+from .lambda_token_pooling import LambdaTokenPooler
